Correctly place the edit (pencil) icons for sections
Adding padding after .mw-headline was a good solution when .edit-page was
positioned absolutely. But absolute positioning caused other problems
(T90245) so .edit-page had to be positioned relatively and floated right.
This solution didn't work well either because .mw-headline didn't have
width specified (couldn't use calc because it's not supported well).

Changing the display of .mw-headline and wrappning .edit-page in a span
and settings its display to table-cell seems to have fixed the problem.

Bug: T105115
Change-Id: I0e407195da36d2f2489c831348b3cc96073ca0b8