Hygiene: Adjust styling of main menu so browser tests do not fail
Confusingly the header-title element covers the entire header. I'm not sure
what purpose this is for, given that we do not want to cover the notifications
icon (and will not due to z-index) and the search icon to the left of it does
the same thing.

This however does confuse browser tests which attempt to click on the left search
icon but cannot click it, as the header accepts the clicks.

If the sole purpose is to center without the notifications icon, given that this is beta
I'd rather remove this now and revisit it in the event we push to beta.

Bug: T105878
Change-Id: I3a1ab6f615f80ee5947dc0b0cbe0112d88a9a309