Use OO.ui button for category overlay and fix saving spinner
The progressive category buttons (which can be added in the CategoryAddOverlay
when searching and clicking one of the suggestions) had a missing check icon. Fix
this by using OO.ui (which is already loaded) to create this button, instead of the Button
class.

Fix: If you click "Save" the header height was increased by a fixed height of the spinner.
Remove the fixed height, because the bug, which introduced it (Bug T74272) doesn't
apply anymore.

Change-Id: I0573641f6fb25cbd71f941fd7fdc02f4baed4966