Always show a logo on Login/Create page
Before this change, the logo was only visible, if there is no error
or warning message. That is really confusing and differs from the
behavior of the core login page.

This change changes the logo to appear everytime the login page is
requested, no matter, if there is an error or warning message.

Follow up: I1b09e7ba80d18dc8c19b2aa475c4d269dffc295e

Bug: T106352
Change-Id: Ib9893b3376c6d15843c2a798a334c1998526e8da