Fix disable Images via Special:MobileOptions
A check, if a checkbox was set, or not doesn't make sense, the browser
normally omit's the checkbox value if unchecked and no special value
given. That means, that an unchecked checkbox for disabled Images would
be interpreted as "Nothing changed, go ahead!" without changing the
value.

Remove this check to enable the disable image setting again and work
around T98187 by explicitly send the current state, if images are disabled
or not, in BetaOptin panel.

Follow up: Ifef83b8fa369f19f9a9d6eb4ab2164271104f8ea

Bug: T106749
Change-Id: I7988eb892397d7430c0f51181d2d6a3af06d804e