Adjust how position-fixed works in light of new beta
Since all drawers/overlays get added to the viewport/body tag they
do not belong to either the content area so it's not possible to distinguish
whether the overlay/drawer should be shifted when the menu is open. Since this
only seems to be a problem for the search overlay - now the only overlay you
can open from the left menu let's limit these rules to drawers for the time
being.

Bug: T106434
Change-Id: Iec5659107f120939efb9e07607bb063ad6ea4012