Update patch set 3
Patch Set 3:

That also seems like a flaw :-) Since this feature is in beta - seems like a good time to fix these to make it easier for the next person to build one.

Seems like margin and height should only be set in certain circumstances (e.g. not in the header)

Patch-set: 3