Diff should be centered in beta
Add post-content class to the user info box - given that it appears
after the content and it is better to use a generic class than a specific
one.

Make the content of the header, menu icons and diff user info line up.

(Note whilst working on this patch I noticed that the last modified bar has
a pre-content and post-content class - this is bad and should be fixed asap -
the pre-content class is misleading)

Bug: T103728
Change-Id: I58ecd8c2d049b468713a6f9f24a278378058d543