Remove new page handling from beta
This seems to be a dead experiment (that I initiated
and got no interest). Now that it is breaking global user pages
in alpha and beta (see bug report) I think it is a good time
to remove it. The purpose was to optimise the non-existent page
for mobile screens but given we have pushed that to stable this
problem should be addressed separately and outside the skin.

Also: Remove out of date comment.

Bug: T102887
Change-Id: Ia9b88cf60f9c6a0834d296b2c818df0a5e200581