Make registered schemas reflect active schemas
... and vice versa.

The list of active schemas [0] was updated to include the
MobileWebBrowse and MobileWebSearch schemas and those schemas and the
MobileWebWatching schema were marked as active. The MobileWebUploads
schema was marked as inactive and removed from the list of registered
schemas.

To make visual diffing a little easier on the eyes, sort the list of
registered schemas alphabetically.

[0] https://meta.wikimedia.org/wiki/Category:Schemas_(active)

Change-Id: I7f3e15a92bd8c340ac319ed141e494b865e94809