Implement a way in MobileContext::getMobileUrl() to work with local URLs
Currently, if local URLs are passed to MobileContext::getMobileUrl(), it will return
an empty string. The underlying funtions to parse the host part and built the
mobile Url expect to get a fully qualified Url.

If parsing the URL with wfParseUrl fails (~false), try to expand the URL to a fully
qualified one and try to parse it again, if it could be expanded. If the Url can be
parsed, use this one to build the mobile url.

Bug: T107505
Change-Id: I092032002f7cb8949f7dc58d7602234764a46d7e