Update patch set 3
Patch Set 3:

@Jdlrobson: That won't work on desktop, testing is possible with chrome and emulating a touch device.

I would say, that we don't change the behaviour of a (long) click on a link. A click should follow the link and a long press should open a context menu to open the link in a new tab. That's really basic browser functionality, which should work on any website. I think it would be really confusing for users if this changes on Wikipedia :)

Patch-set: 3