Update patch set 67
Patch Set 67:

I'd really like to find a way to move this forward as it's been sitting here on the horizon for a while now. The main issue seems to be around documentation so let's find a way to work that forward. I've pinged Bryan on the task to try and work out what our options are here. Nothing feels satisfactory right now..

Patch-set: 67