Update patch set 5
Patch Set 5:

No plans to ever merge this Paladox for the reasons given on the bug T65459. 

I would really appreciate your help on https://gerrit.wikimedia.org/r/229728 which seems like a good approach which would give site admins the ability to customise the menu and add this link and thus solve the problem.

I just wanted to set that expectation, in case you were expecting a review.

Patch-set: 5