Update patch set 70
Patch Set 70:

Kaldari - in theory we could find a way to convert it to a README, this is why I figure as close to the code as possible is better. we could change @doc keys to @doc:key @doc:key2 @doc:key3 etc... and then easily generate a mediawiki document per release. Much more likely to be kept up to date.

Patch-set: 70