Fix toolbar position in iOS
When the anything is focused in iOS it calculates the overlap between
the desired cursor position and the keyboard, and it scrolls the
window up by that amount.

The problem is the window contains just the toolbar and a scrollable
container, so we end up losing the toolbar.

To work around this, whenever the *window* is scrolled apply
the window's scroll offset to the scrollable container and
reset the window scroll offset to 0.

We also need to reapply the native selection so the cursor doesn't
magically disappear.

Change-Id: Ib4ff64589d299cac53a7d12f2a97dfc3149ea65a