Update patch set 70
Patch Set 70:

The point why I prefer a README is, that we can reflect multi-line comments much easier. We have some complex configuration option descriptions, which has code examples which takes several lines, just an example:

> Make the logos configurable.
>
> Currently, 'copyright', 'copyright-width', and 'copyright-> height' elements are supported.
> 'copyright' is the URL of the logo displayed in the footer
> 'copyright-width' (optional) is the width in pixels of the > copyright image you want to display
> 'copyright-height' (optional) is the height in pixels of 
> the copyright image you want to display
> If the actual 'copyright' dimensions are 200x30, then you > may want to set the width and height
> to 100 and 15 respectively (in order to support retina
> screens).
>
> Example: array(
> 	'copyright' => '/images/mysite_copyright_logo.png',
> 	'copyright-width' => 100,
> 	'copyright-height' => 15,
> );

How do we plan to reflect them in extension.json? In a single line seems pretty useless for developers and wiki admins. We could work around this with a solution like:

"@doc:wgMFCustomLogos": [
	"Make the logos configurable.",
	"",
	"Currently, 'copyright', 'copyright-width', and 'copyright-height' elements are supported.",
	"'copyright' is the URL of the logo displayed in the footer",
	"'copyright-width' (optional) is the width in pixels of the copyright image you want to display",
	"'copyright-height' (optional) is the height in pixels of the copyright image you want to display",
	"If the actual 'copyright' dimensions are 200x30, then you may want to set the width and height",
	"to 100 and 15 respectively (in order to support retina screens).",
	"",
	"Example: array(",
	"  'copyright' => '/images/mysite_copyright_logo.png',",
	"  'copyright-width' => 100,",
	"  'copyright-height' => 15,",
	");"
]

But the question is: Is it a way we want/can handle it? I would like to have a global specification for that. I'm pretty sure, we aren't the first, who run in this problem. :) So a good way would be to work on a global code style for extension.json (and skin.json) especially for documentation of configuration values so that we have an overall design on all extensions.

Patch-set: 70
Label: Verified=0