Update patch set 70
Patch Set 70:

Or maybe we can do something similar to the .jshintrc file which is a json file but can do comments.

But yes we should keep it in the readme file to make it look nice there is also documentation on the mw page.

We can add a notice in extension.json to make sure it is kept updated.

Patch-set: 70