Don't render a toast on page load
Separating init functions of toast from the ToastDrawer itself. ToastDrawer
is the "old" toast.js, which lives in the new module ToastDrawer. toast.js
is a wrapper around one (unique) ToastDrawer instance, which will be created,
when the toast drawer is used the first time, instead of when the toast object
is created.

toast.js wraps all needed and all used Api's of the now called ToastDrawer,
so there is no need to change any existing code. Toast.js (from an Api point
of view) feels like the current implementation.

To save a pending toast, you can use the same function like before, which is
handled directly in the toast module, instead of in the drawer itself. That's
allows us to create a toast out of _showPending() only, if there is a pending
toast message to show. That avoids to render a toast on every page load, even,
if it isn't used directly after or during the page load.

Bug: T108249
Change-Id: Ia4649843f9810007454bdf27612f375c0eea7c5c