Revert "Don't hardcode script tags. Use RL makeInlineScript method"
This reverts commit 75f4cf09a27c019b8ee9e029fa9f3e367699a42a
We should be careful about changing the page HTML and shouldn't
rush merge things. Yes the bug we are fixing is a regression but
it's not a high impact one.

The delay is unacceptable and if we decide this is the way forward
there is no reason to have a `mobile.head` module at all.
Change-Id: I386a6c99cd6700f5927857a6636555c82c9dab2c