Update patch set 2
Patch Set 2:

That would require, that every user of the toast module initialize it's own object of toast? That could result in multiple toasts added to the DOM and maybe overlap each other :/So there should be one global toast which is not added to the DOM before it is used the first time. In this case I like my approach, by I have written it, so.. ;) How would you handle it?

Patch-set: 2