Update patch set 2
Patch Set 2:

The more I think about this, I'm not convinced rendering this is causing any performance issues.

The code is already available so the only cost here is rendering it and it renders as invisible. Is this causing a reflow or repaint? if yes then it matters. If not.. I don't think it does.

Given the size of this change, it seems unnecessary unless there is good means to. Please try convincing me on the task because right now I feel like we should decline this task.

Patch-set: 2