Update patch set 2
Patch Set 2:

It maybe a micro optimisation, but these things build up. we need gain maximum performance if we can. Once the toast is inserted into the DOM, a reflow will happen. Also subsequent querying of the DOM will take longer because there are more elements in it now.

Patch-set: 2