Update patch set 3
Patch Set 3:

@Florianschmidtwelzow I hear you and I'm not saying we should push this to stable. It's clear the swipe isn't discoverable and that's not working hijacking clicks seems the best option right now.

An alternative might be to add an icon to the left of internal links that when clicked opens a preview. Another alternative is to set a user preference to turn off.

Right now the feature imo is not discoverable enough to even warrant being in beta and I'm not convinced swipe is the right decision and am not comfortable with continuing on it in the following form as it seems broken.

Maybe we should put this in a feature branch until we work out a way forward so this doesn't block turning off alpha?

Patch-set: 3