Update patch set 3
Patch Set 3:

Florian thanks for pushing this forward. Come to think about it, it maybe hard for some users to swipe over links (as it is on iOS, presumably). I like the idea of making it a simple config option that is stored in local storage. Or just moving the code as is to the feature branch as you guys said.

Patch-set: 3
Label: Verified=0