Update patch set 4
Patch Set 4: Code-Review+1

I separated the move to beta step so from my perspective this is good to merge.

It's worth pointing out if we hijack links, you'd still be able to long press and open the page in a new tab or new window. I agree some people might not like it though (that said there are lots of people that don't like mobile section collapsing because it stops you finding and replacing) - it could be a good thing to A/B test. Do people interact more with links with this feature enabled?

So let's merge this then do a follow up to either
* remove and branch it
* switch it to work on press

Patch-set: 4
Label: Code-Review=+1