hygiene: Goodbye custom event emitter and class code
This patch is not as scary as it looks.
Best starting point is noting the removal of eventemitter and the changes
in Class.js, the rest just reflect these changes.

Changes:
* Introduce OO.mfExtend - a function that hopefully we can upstream to OOJS but
in mean time helps us with adoption.
* Remove EventEmitter code
* All classes previously inheriting from EventEmitter now mix it in
* ModuleLoader in mobile.modules now depends on OO. OOJS now loads
before it
* Remove _parent magic: Previously a View magically inherited defaults and templatePartials from
its parent. This was a little confusing and broken, as it only worked for
the immediate parent. Let's remove this code and bring ourselves a little
closer to oojs! This now means you must use $.extend explicitly on templatePartials and
defaults
* extend function is now moved from Class to View. Will be deprecated later to make
us more consistent.
* Class is left in repository given that there are various usages in places such as Gather.

Depends-On:  Icaf2d24c2764658fa775600965529f1a609262b6
Change-Id: I5374b2384b1e464cc5312b95bb482ed79f1df70e