Update patch set 3
Patch Set 3: Code-Review+1

I'd like someone else to try and break this too.

This LGTM. G as in Great!

I've noticed yet another shift in the position of the content when you open the search overlay (to the left this time), but this is present in stable in master too.

Patch-set: 3
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1