Update patch set 1
Patch Set 1: Code-Review-1

Thanks, and apologies for the confusion on the thread - I hadn't fully understood that you literally meant use the addModuleStyles/Scripts functions (ah the maze of RL) and now I'm trying to work out how it knows to conditionally add it.

I still think our efforts would be better spent identifying why this code needs to exist and remove it as otherwise it's likely to just stay here and this just prolongs that.

That said the styles need to declare there position to avoid warnings in production so there is an issue with this patch in current form:
`Warning: OutputPage::getModuleStyles: style module should define its position explicitly: mobile.usermodule MobileUserModule `
(it should be position bottom)

Patch-set: 1
Label: Code-Review=-1