Hide languagelinks list on alpha/beta login page
The list is a list of links now and can be very big for some wikis, which
makes the login page (especially on mobile phones) very hard to read, understand
and use (especially for new users). Hide it for now, until we have a much nicer
way to do this.

Bug: T110403
Change-Id: If62f8dc1d97ae8d2b6bb456394fd44714f4daecb