Update patch set 16
Patch Set 16:

I would really appreciate a review. It's existing in many forms over the past few months :)

I have a few oojs ui related activities in a local github that are dependent on this work and I'm keen to get this pushed through before continuing.

It's also going to help with efforts to upstream Gather and other mobile code to desktop.

Patch-set: 16
Label: Verified=0