Update patch set 3
Patch Set 3:

@Jdlrobson See my comment here: https://phabricator.wikimedia.org/T85368

We plan to use Cirrus for search on Wikidata in future but for now it would be the easiest solution to just hook into the javascript code. When we finished the migration to Cirrus, we will be able to use the default mediawiki search like mobile does anyways and the javascript hack will be obsolete. Are you ok with that solution?

Patch-set: 3
Reviewer: Bene <549@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0