Update patch set 7
Patch Set 7:

I've corrected the event type for the logger, but this is a little confusing; we call the mw.echo.Logger 4 times in the code, three for impressions and one for click, which doesn't sound right to me. Then again, the code seems to be correct in terms of where it's placed.

Can you double check?

The four instances are:

resources/mobile.notifications.overlay/NotificationsOverlay.js

- notificationClick, Line #92 (inside on( 'click'...))
- notificationImpression, Line #103 (for each '.mw-echo-notification' initialization)

resources/mobile.special.notifications.scripts/notifications.js
- notificationImpression, Line #31 (for each '.mw-echo-notification' initialization)
- notificationImpression, Line #90 (for each '.mw-echo-notification' initialization that loads in 'read more')

There's no notificationClick for each *click* of the notifications on the special page, because they're links and don't have a specific "click" event listener. 

Should I add that? We should consider/reassess our use of events and where anyways at some point anyways, so this can be done in a followup commit once we decided if this should be done in general in Echo regardless of media (mobile/desktop)

Patch-set: 7
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0