Update patch set 11
Patch Set 11: Code-Review-1

(2 comments)

Bmansurov I think it's okay to only show a placeholder with JS disabled given that with JS disabled you see a search button which gives it context.

I'm not too happy with the class introduction/removal of mw-ui-input though and I would like some more input on why about that.

In terms of a CSS solution I think we can iterate on this, so no need to go down that route just yet.

Patch-set: 11
Label: Code-Review=-1