Let interested parties know whether the beta opt-in panel is shown
Keep track of the "minerva.betaoptin" topic using mw.track. The
interested parties will be able to subscribe to the topic using
mw.trackSubscribe. The solution also works well regardless of
the load order of the "skins.minerva.scripts" module.

This helps other extensions act accordingly depending on whether
the beta opt-in panel is shown or hidden.

Bug: T111188
Change-Id: I644da82a6e1d7dea8e6047a60bdefb14982ac010