Update patch set 14
Patch Set 14:

(1 comment)

I'm really uncomfortable with having an input field withput any border or any other indication, that there is an input field. With the placeholder we had something that indicates, that there is an input field for the search query, not we have a white input on a white background -> not really good. For me, that would be a deterioration for non-JS users :(

@Jdlrobson: What do you mean?

Patch-set: 14