Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

One minor issue - the main page has a styling regression http://imgur.com/SmSCfdk,9aGAzgq

I've checked cached pages and they are fine.

Also we shouldn't merge this until the max width has been applied to (since the bar was not designed to be at the bottom without it) so I would suggest we do this work in a dev branch and then squash the changes and push to master in one commit. Thoughts ?

Patch-set: 3
Label: Code-Review=-1