Update patch set 21
Patch Set 21: Code-Review-1

This change entails a few distinct changes that I don't see mentioned in the commit message.

I'm happy to see dependencies being made explicit (the pageApi-related changes) – we can test more code in isolation (!) – but it's not clear to me why they are necessary in this patch from the description given.

What I mean to say is, please provide more context for this change. State and explain the problem is and then detail your solution.

Patch-set: 21