Update patch set 2
Patch Set 2:

My -1 still stands. We are add 100 lines of code, for a micro optimisation that might be outweighed by the addition of code size and may introduce unforeseen bugs.

This also refactors a very important part of our infrastructure - our ability to send notifications. Is this really worth it?

Please ping Rob Moen or Sam Smith. If you can convince either of them (or indeed me) this is worthwhile I'll retract my -1.

Patch-set: 2