Update patch set 6
Patch Set 6: Code-Review+2

We discussed this yesterday during standup. It might increase the amount of time taken to replace the last modified bar on cached pages but we're accepting that as a cost of simplifying everything – we're also moving the last modified bar to the bottom of the page at the same time so it won't be an issue once the page has been edited.

Patch-set: 6
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2