Rename modules in resources
Resource files have been moved to their own "resource" folder, but the
module names have not been modified to reflect this. Therefore, the
new names bring it back to the old convention of identifying the module
name based on the path.
The old module names have not been removed, and have been deprecated so
that other extensions do not break.

Bug: T100467
Change-Id: I07b43666db565183554012bf5eae46a3df911735