Update patch set 1
Patch Set 1:

(1 comment)

We don't have a case where .printonly class is used by itself. Applying it to every .printonly (which may not exist outside citations) maybe a premature optimization and may have side effects that I'm unaware of.

Yes it should appear when printing. By default it will appear on any medium, but as you can see in the code I'm hiding it on certain media which doesn't include printers.

Patch-set: 1