Update patch set 10
Patch Set 10:

With a simple comment "recheck" :)

Unfortunately, Jenkins run all tests for you, if you upload a new patchset. That means, that all tests, that would be triggered with a recheck comment from you, already run, when you upload a new patchset. For some other users (like me and Jon), Jenkins will run the "extended" test suite for all changes (and all changes we trigger with recheck). The reason is, that the tests take a long time to run and with that we can minimize the run on changes, that maybe doesn't follow our coding standards or some other basic test :) That's how I would explain it :P

Patch-set: 10
Label: Verified=0