Update patch set 8
Patch Set 8:

We found a solution which does not need to replace the whole api call but only needs configuration to replace the generator used to create the search results. The hook therefore is unneeded and this patch should be abandoned if it isn't useful anywhere else.

I will publish my drafts as soon as we settled on a name for the new generator api module. You can find the relevant patch here: https://gerrit.wikimedia.org/r/#/c/238771/

Patch-set: 8