Update patch set 12
Patch Set 12: Code-Review-1

(9 comments)

I tested this locally and found the issues inline.

With those issues fixed, I've been able to create a new collection in Gather and search for pages to add to it; search in stable and beta; view a page's categories and add that page to new categories.

Patch-set: 12
Label: Code-Review=-1