Fix media viewer for files with special characters
The file's title is never encoded before it's added to the URL fragment
despite the URL fragment being decoded prior to rendering the media
viewer. When the file's title contains special characters, then the
decode step throws an error in Safari 8 (and possibly others).

Encode the file's title before it's added to the URL fragment.

Bug: T111609
Change-Id: Ib09b791c9a01868a04eb8d079eb4a7704dadeaad