Update patch set 2
Patch Set 2:

> I think I understand your problem, but wouldn't it work, if you re-use the MobileLicenseLink hook for it? If not, I'm fine with merging this :)

Sadly not, as MobileFrontend wraps the message you give it in some text which doesn't really fit with what we have on Wikidata. For English we could probably create a new Wikidata specific message that would fit into this sentence, but that would be oddly worded. For other languages this might not work at all.

Patch-set: 2
Label: Verified=0