Update patch set 3
Patch Set 3:

I agree, that this isn't the best solution. But the hook isn't documented, and I wouldn't advise anyone to use it, because we will not invest a lot of time to maintain it (nearly zero).

The other hand is (it sounds more badly in a pure text form, as it is meant): do you plan to rework this part of MobileFrontend in the next time? Does the reading web team plan to invest work/time into this function this quarter or the next? I think no, which is why I think that's the best solution until we conpletly rework the license code part.

Patch-set: 3