Breaking change: Move logic for section wrapping out of skin into hook
This allows other skins to benefit from having lead sections marked up.
This is backwards compatible with old HTML for sections (checked vigorously)

Changes:
* Update toggling code to only take into account children so that table of
contents is not included if it has rendered already
* #content now has .mw-body class so that notification area is not added inside
the content element (see mw.util)

Breaking change:
* Now entire content is wrapped in element with id mw-content-text - this
is consistent with desktop but now means the lead section is one more deep
then before. For clients using Page.getLeadSection this will not be a problem
but things like QuickSurveys will need to be updated. Given the deployment
freeze this is a good time to do this! Please can merger open a bug to fix
QuickSurveys in mobile?

Bug: T115147
Change-Id: I11fd44f183a95c7564961c6273235c126eaa3b8b