Update patch set 9
Patch Set 9:

Hi I haven't ignored them. I updated the comment msg to make it easer to others what the problem was when using extension.json.

-> Please don't merge these modules. Keep $wgMinervaStyleModules and do a $wgResourceModules+=$wgMinervaStyleModules at the bottom of the file.

Do you mean change it so that it dosent merge three like this for example

wgResourceModules = array_merge( $wgResourceModules,

So it is not using array_merge.

And could I ask why carnt we do it this way. Is there anything stopping it from doing it this way. will it brake other things.

Patch-set: 9