Update patch set 9
Patch Set 9:

You can still use array_merge, I would just ask that you minimise the unnecessary changes so the diff is more readable.
For instance because of the way you've gone about this it looks like $wgMobileEchoModules has been added when it was already there.


The variables $wgMinervaStyleModules, wgMinervaBootstrapModules and -$wgMobileSpecialPageModule were serving purposes.

They help guide the developer what belongs in there and give clues to developers when added to a skin to check they are being added in the right place.

The same patch could be written with a line at the bottom
+$wgResourceModules += $wgMobileModules + $wgMobileSpecialPageModules ...

Patch-set: 9