Remove 'header-loaded' event
Previously we were registering the click events of the
main menu button once the header loaded. This doesn't
make much sense since the menu might not have been
added to the page yet and the menu could appear blank
on slower connections, which could confuse users.

Given that the code now runs when the DOM is ready, it
makes sense to move this to the bottom of the page.

Changes:
 * Remove 'header-loaded' event
 * Correct module name ('mobile.head' no longer exists)
 * Remove module skins.minverva.scripts.top

Bug: T112315
Change-Id: Id73dfbb523bde91cca136f90b5e1ce825343b1bc