Update patch set 1
Patch Set 1:

Why do you say that? Based on what information? It's not just "nasty HTML". There is a clear specification of the very strict rules applied on displaytitle, and the fact that the Title object used in this code actually supports (!) displaytitle tells me that this is fine. So what is the reason for the -1? What are you missing? What should I improve?

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0