Update patch set 2
Patch Set 2: Code-Review-1

eww, I'm feeling bad putting a workaround into MobileFrontend, which is just needed for wikidata :/ Couldn't you "just"(tm) hook into the search api module and set the label of the element as the displaytitle, which would work, if I37fa6b6d58 is merged? If this isn't possible right now, I would like to investigate, why and maybe implement a way of doing this :) That would allow other extensions to change the displaytitle, too (if someone wants to do that :P).

Patch-set: 2
Label: Code-Review=-1