Update patch set 1
Patch Set 1:

@florian stuff like the formatting in the user page exists:

https://en.wikipedia.org/wiki/User:Levdr1lostpassword

this mostly (almost always?) happens only in user space and non-articles, so maybe this is okay enough.

some of the formatting is good, like italics for titles or formatting elements with subscripts.

maybe we can override color formatting (and maybe size changes?) in special nearby to work around any issues

Patch-set: 1