Update patch set 3
Patch Set 3:

@Thiemo Mättig (WMDE): Couldn't you support the displaytitle page prop? That's what the NearbyGateway currently uses (through the pageprops api module). If not (e.g. because of Localisation), maybe we can add a hook to the pageprops api module[1] to support the change of pageprops returned by the database. For performance reason there is maybe a way to work around the initial db query to get all pageprops of a page so that you can define them in Wikibase. (I'm not very experienced in how the api works internally, but I would be happy to assist as much as I can).

What I want to say: If you add a workaround for Wikibase in MobileFrontend it will work for MF, but what if there are other clients of MW and WB in the future? You always need to add a workaround in the interface, which seems false, if we could work for a solution which lives on the server side. In my opinion, the ideal solution would be, if a query to the displaytitle page prop always returns the real display title, no matter, if it's set by WB, the DISPLAYTITLE parser function or any other extension :)

[1] https://github.com/wikimedia/mediawiki/blob/9160785c6fb478a929502372ce00a0560f0de936/includes/api/ApiQueryPageProps.php#L97-L100

Patch-set: 3