Update patch set 3
Patch Set 3:

I haven't fully read your comment before writing mine. Sorry for that.

Wikibase can not simply store a displaytitle pageprop, because Wikibase is multilingual but the pareprops are not.

A hook could work. We will investigate. Thanks for the suggestion.

However, this patch is supposed to be a temporary workaround around an pressing issue that can very easily be solved that way (as you can see). We would like to have this feature in a usable state next week, before the third Wikidata birthday. We will continue working on this, as we always do.

I'm adding Wikidata's product manager Lydia Pintscher.

Patch-set: 3