Update patch set 3
Patch Set 3:

If you add a FIXME notice to the qiery explaining the problem and linking to a task I wpuld be ok with introducing this now. But the task should contain the problem iself and I would like to see a commitment that you will fix the problem in the next time, say 2-3 month?

Patch-set: 3
Label: Verified=0