Update patch set 3
Patch Set 3: Code-Review-1

The fact that display title cannot be used is technical debt in wikidata that I do not wish for MobileFrontend to inherit. We are in the process of breaking apart MobileFrontend into minimal components and this kind of thing could derail that.

We can simply not have lines of code that purely serve wikidata. We have a responsibility as developers to find and define interfaces that make sense for our code for it to be future proof. No more hacks please.

I'm definitely open to working out architecture changes in mediawiki itself to support this usecase and I'm open to hooks/ wikidata using and tweaking our code as we did previously with the search api (if you want hacks in wikidata that's your choice)

I would throw down a -2 here normally but I trust that you will pay attention to this discussion and work to find a better solution.

Patch-set: 3
Label: Code-Review=-1