Update patch set 5
Patch Set 5: Code-Review-2

> @Jdlrobson Yes, a multilingual display title in core would be awesome. But it doesn't fit with how core does things currently, and changing that opens the entire "how to manage multilingual content" can of worms. That's a long term re-architecturing goal. We need something that we can roll out next week.

Although I can sympathise with wanting to release fast, and my team has done this historically, we are trying to reverse this trend as we see the dangers of this approach. Because of this I ask that you please don't do these kind of hacks inside Mobilefrontend. If you want to take that approach your code needs to be contained in Wikidata.  
I do not want MobileFrontend to inherit code debt because of Wikidata - we have enough of our own. This doesn't fit in with our long term plan for MobileFrontend which is that this doesn't exist.

Please do start this re-start this conversation about multilingual support in core on wikitech/phabricator. This attitude is what stops us from actually ever making big architectural changes which hurts us all. I would be keen to help work with you and drive those changes. Plus when making these changes it's always useful to document in the code that this is the task that is unresolved that is forcing us to do that.

Everytime MobileFrontend implements something you're gonna have to monkey patch it - search,nearby.. what's next? That's not sustainable.

Something worth exploring is we have a Page JavaScript class that all these things use - if we can get that to behave how you want it too everything should just magically work - e.g. add a getTitle method https://doc.wikimedia.org/MobileFrontend/master/js/#!/api/Page

Given all these changes you are attempting to do in our codebase, I'm putting down a -2 until we discuss this further - I would suggest on the mobile-l mailing list pointing to relevant bugs - as it would be wise to talk through changes and agree on them before implementing. You also have my teams attention if you join us in #wikimedia-mobile to chat there.

It would also be great to hear from your team on the mobile-l mailing list that you are working on making MobileFrontend mobile friendly and what problems you have help you need - springing this on us is a huge surprise (at least to me) and I want to help but I think we need to take time to think through the architecture. You may want a quick fix now, but let's save all the future bugs/regressions that are going to hit you by taking such an approach.

Patch-set: 5
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2