Update patch set 5
Patch Set 5:

Aude I agree that your approach in the other patchset looks sounds and yes please do start that conversation - poking relevant RFC/phabricator tasks and creating new ones where needed.

In terms of short term I've suggested a more sane approach which allows MobileFrontend to continue to be agnostic to whether Wikidata exists or not, and improves MobileFrontend in the process. It needs a bit of work though:
https://gerrit.wikimedia.org/r/246282

Patch-set: 5