Update patch set 21
Patch Set 21:

But it works.

Should I use array_merge or not.

It works.

But this way seems to be the nice way since they are all using different resoloader modules settings that are added at the bottom and includes a new resourceloader module settings since it replaces the array_merge $wgResouceLoader code.

And does it as if it was just all together.

Patch-set: 21