Update patch set 21
Patch Set 21:

> But it works.
>Should I use array_merge or not.
> It works.

I'm not sure how much clearer to be.. I don't /care/ if it works. I want to understand why. As a programmer you should be caring about these things and understanding the underlying system.

If you were given the problem "there is a smell of gas in my house", covering your nose or cutting off your nose could be solutions to that problem but they might not necessarily be the right solution.

Does this make sense?

Patch-set: 21