Hygiene: Introduce Toggler class, kill toggle instance
This is going to be needed to be able to emit events on the opening
and collapsing on sections but not in table of contents.

It also allows us to detach toggling in
table of contents from the sections in the long term.

Changes:
* Breaking change for anything using the toggle module (there should be
none outside MobileFrontend)
* Minor code refactor
* expandStoredSections now takes a Toggler instance (breaking change
but a private function so shouldn't be being used anyway)

Change-Id: I38c17eacc452942aea8942328c7b26cff0e50d99